+++ src/app/Pages/provider/provider.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Subscription } from 'rxjs';
-import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { NgbModal, ModalDismissReasons } from '@ng-bootstrap/ng-bootstrap';
 import { CommercesService } from 'src/app/Services/commerces.service';
 import { Router } from '@angular/router';
 import { viewAttached } from '@angular/core/src/render3/instructions';
@@ -25,11 +25,14 @@ export class ProviderComponent implements OnInit {
 
   public providers:any;
   private commerceSubscription: Subscription;
+  providerValue;
+  closeResult: string;
   
   constructor(
     public _providersService:ProvidersService,
     public router: Router,
-    private toastr: ToastrService
+    private toastr: ToastrService,
+    private modalService: NgbModal
   ) {  
     this.providers = "";
   }
@@ -58,20 +61,43 @@ export class ProviderComponent implements OnInit {
 
   UpdateService(provider){
 
-   
-
+  
     
   }
 
-  deleteService(provider){
-    this._providersService.deleteProvider(provider).subscribe(
-      response=>{
-        this.toastr.info(provider.name+' ha sido borrado!','Servicio Borrado', {
-          timeOut: 5000,
-        });
-        this.obtenerProveedores();
+  deleteProvider(content,provider,$event){
+    $event.stopPropagation();
+    this.modalService.open(content).result.then((result) => {
+      this.closeResult = `Closed with: ${result}`;
+      if(result == "si"){
+        this._providersService.deleteProvider(provider).subscribe(
+          response=>{
+            this.toastr.info(provider.name+' ha sido borrado!','Servicio Borrado', {
+              timeOut: 5000,
+            });
+            this.obtenerProveedores();
+          }
+        )
       }
-    )
+    }, (reason) => {
+      this.closeResult = `Dismissed ${this.getDismissReason(reason)}`;
+    });
+
   }
+
+  public updateTable(){
+    this.providerValue="";
+  }
+
+  private getDismissReason(reason: any): string {
+    if (reason === ModalDismissReasons.ESC) {
+      return 'by pressing ESC';
+    } else if (reason === ModalDismissReasons.BACKDROP_CLICK) {
+      return 'by clicking on a backdrop';
+    } else {
+      return `with: ${reason}`;
+    }
+  }
+
 }
